requirements.txt
opencv-python
numpy
ultralytics
python-dotenv
requests